Galaxy Dynamics of NGC2403 and its Dark Matter Content
In this project we will learn about this mysterious
dark matter halo and its properties in the local Universe.

This project is aimed to teach the chronological history of dark mat-
ter, galaxy dynamics, and determination of dark matter content in galaxies

(theoretical+observational). We will use tools such as python, matplotlib,
and topcat for reading, writing, and visualizing the datasets. In particular,
we pick-up a local galaxy named NGC2403, and determine its dark matter
content using robust techniques.